Refactor: Extract useMarkdown hook & MarkdownPreview component
